Updating task 4 to remove decimal point

+++ supervised_learning/0x00-binary_classification/4-neuron.py
@@ -48,4 +48,4 @@ class Neuron():
         """evaluates neuron's predictions"""
         A = self.forward_prop(X)
         C = self.cost(Y, A)
-        return np.round(A), C
+        return np.round(A).astype(int), C
